add check for updates button
