+++ WolvenKit/Views/HomePage/Pages/SettingsPageView.xaml
@@ -27,25 +27,25 @@
         <syncfusion:PropertyGrid x:Name="SettingsPropertygrid" Grid.Row="0"  Margin="3,75,3,0"
                             AutoGeneratingPropertyGridItem="SettingsPropertygrid_OnAutoGeneratingPropertyGridItem"
                             CategoryForeground="{DynamicResource MahApps.Brushes.Accent}"
-                            EnableGrouping="True"
-                            SelectedObject="{Binding Settings, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
+                            EnableGrouping="True"/>
 
         <Grid Grid.Row="1">
 
             <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="Auto"/>
                 <ColumnDefinition Width="*"/>
                 <ColumnDefinition Width="Auto"/>
             </Grid.ColumnDefinitions>
-            
-            <StackPanel Grid.Column="1" Margin="5" HorizontalAlignment="Right" Orientation="Horizontal" >
-                <TextBlock Margin="5"
-                        HorizontalAlignment="Center"
+
+            <StackPanel Grid.Column="0" Margin="5" HorizontalAlignment="Right" Orientation="Horizontal" >
+                <Button x:Name="CheckForUpdatesButton" Content="Check for updates" />
+            </StackPanel>
+
+            <StackPanel Grid.Column="2" Margin="5" HorizontalAlignment="Right" Orientation="Horizontal" >
+                <TextBlock Margin="5" HorizontalAlignment="Center" VerticalAlignment="Center"
                         Foreground="{DynamicResource MahApps.Brushes.AccentBase}"
                         Text="Theme changes will take affect after restart." />
-                <Button
-                x:Name="ExitRestart"
-                Click="ExitRestart_Click"
-                Content="Exit and Restart" />
+                <Button x:Name="ExitRestart" Click="ExitRestart_Click" Content="Exit and restart" />
 
             </StackPanel>
         </Grid>
