+++ WolvenKit/Views/HomePage/Pages/SettingsPageView.xaml.cs
@@ -1,17 +1,12 @@
-using System.Reflection;
+using System.Reactive.Disposables;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
 using ReactiveUI;
 using Splat;
-using System.Windows.Media;
-using Syncfusion.Windows.Controls.Layout;
 using Syncfusion.Windows.PropertyGrid;
-using Syncfusion.Windows.Tools.Controls;
 using WolvenKit.Functionality.Services;
 using WolvenKit.ViewModels;
-using System;
-using System.Globalization;
+using static WolvenKit.Converters.PropertyGridEditors;
 
 namespace WolvenKit.Views.HomePage.Pages
 {
@@ -25,6 +20,19 @@ namespace WolvenKit.Views.HomePage.Pages
 
             ViewModel = Locator.Current.GetService<SettingsPageViewModel>();
             DataContext = ViewModel;
+
+            this.WhenActivated(disposables =>
+            {
+                this.Bind(ViewModel,
+                    viewModel => viewModel.Settings,
+                    view => view.SettingsPropertygrid.SelectedObject)
+                .DisposeWith(disposables);
+
+                this.BindCommand(ViewModel,
+                      viewModel => viewModel.CheckForUpdatesCommand,
+                      view => view.CheckForUpdatesButton)
+                .DisposeWith(disposables);
+            });
         }
 
         #endregion Constructors
@@ -67,89 +75,7 @@ namespace WolvenKit.Views.HomePage.Pages
                         propertyItem.Editor = new BrushEditor();
                         break;
                 }
-
-                
             }
         }
-
-        internal class BrushEditor : ITypeEditor
-        {
-            private ColorPickerPalette _editor;
-
-            public void Attach(PropertyViewItem property, PropertyItem info)
-            {
-                if (info.CanWrite)
-                {
-                    var binding = new Binding("Value")
-                    {
-                        Mode = BindingMode.TwoWay,
-                        Source = info,
-                        ValidatesOnExceptions = true,
-                        ValidatesOnDataErrors = true,
-                        Converter = new StringColorConverter()
-                    };
-                    BindingOperations.SetBinding(_editor, ColorPickerPalette.ColorProperty, binding);
-                }
-                else
-                {
-                    _editor.SetCurrentValue(UIElement.IsEnabledProperty, false);
-                    var binding = new Binding("Value")
-                    {
-                        Source = info,
-                        ValidatesOnExceptions = true,
-                        ValidatesOnDataErrors = true
-                    };
-                    BindingOperations.SetBinding(_editor, ColorPickerPalette.ColorProperty, binding);
-                }
-            }
-            public object Create(PropertyInfo propertyInfo)
-            {
-                _editor = new ColorPickerPalette();
-
-                return _editor;
-            }
-            public void Detach(PropertyViewItem property)
-            {
-
-            }
-        }
-
-        [ValueConversion(typeof(Color), typeof(String))]
-        public sealed class StringColorConverter : IValueConverter
-        {
-            public static readonly StringColorConverter Default = new StringColorConverter();
-
-            // convert color to string
-            public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-            {
-                if (value is string strValue)
-                {
-                    try
-                    {
-                        var color = (Color)ColorConverter.ConvertFromString(strValue);
-                        return color;
-                    }
-                    catch (FormatException)
-                    {
-                        return DependencyProperty.UnsetValue;
-                    }
-                }
-                return DependencyProperty.UnsetValue;
-            }
-
-            // convert string to color or unsetvalue
-            public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-            {
-                if (value is Color color)
-                {
-                    return color.ToString();
-                }
-                return DependencyProperty.UnsetValue;
-            }
-        }
-
     }
-
-
-    
 }
